hpcgap/src/opers.c: sync
